pyproject.toml
[tool.poetry]
name = "fastapi-tdd"
version = "0.1.0"
description = ""
authors = ["lemondouble <lemondouble2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
httpx = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pytest-alembic = "^0.11.1"
docker = "^7.1.0"
faker = "^28.1.0"
ruff = "^0.6.4"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
